First version
This represents the initial release, constituting the first version of the project. Included within this version are illustrative examples designed to demonstrate its functionality and usage. Comprehensive documentation is also provided as part of this delivery, offering guidance and details. The capabilities integrated into this specific version encompass the implementation of Taylor regression, alongside the support for Piecewise Taylor regression.